insert data in recyclerview
