--- app/src/main/java/org/sopt/sample/ui/main/HomeFragment.kt
@@ -1,40 +1,0 @@
-package org.sopt.sample.ui.main
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import org.sopt.sample.databinding.FragmentHomeBinding
-
-
-class HomeFragment : Fragment() {
-    private var _binding: FragmentHomeBinding? = null
-    private val binding get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 바인딩 초기화 에러가 발생했습니다." }
-    private val adapter by lazy { SampleAdapter() }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentHomeBinding.inflate(layoutInflater)
-        initAdapter()
-        setList()
-        return binding.root
-    }
-
-    private fun initAdapter(){
-        binding.recyclerViewHome.adapter = adapter
-    }
-
-    private fun setList(){
-        val list : List<String> = listOf("1", "2", "3", "4", "5", "6", "7", "8", "9", "10")
-        adapter.sampleList = list
-        adapter.notifyDataSetChanged()
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
